// store-flow-prototype-font-end-angular/src/app/core/models/user.model.ts
export interface UserInfo {
    id?: string | null;
    username: string | null;
    nameTH?: string | null;
    surnameTH?: string | null;
    nameEN?: string | null;
    surnameEN?: string | null;
    telephone?: string | null;
    mobilePhone?: string | null;
    idCard?: string | null;
    email?: string | null;
    brithDate?: string | null;
    roleNameTH?: string[] | null;
    roleNameEN?: string[] | null;
}